+++ public/admin/_app/immutable/nodes/1.CgWJePl7.js
@@ -1,0 +1,1 @@
+import{$ as e,Ct as t,F as n,L as r,N as i,at as a,it as o,jt as s,ot as c,wt as l}from"../chunks/Bny3TPnM.js";import"../chunks/DYR6MKK2.js";import{t as u}from"../chunks/CUVWf82p.js";var d=r(`<h1> </h1> <p> </p>`,1);function f(r,f){l(f,!0);var p=d(),m=a(p),h=o(m,!0);s(m);var g=c(m,2),_=o(g,!0);s(g),e(()=>{i(h,u.status),i(_,u.error?.message)}),n(r,p),t()}export{f as component};