+++ public/admin/_app/immutable/nodes/24.DXiPdcYA.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/B-b4oPOB.js";export{e as component};